feat: 煎药室管理模块开发完成，通过测试

+++ dms-app/src/main/resources/db/migration/V16__fix_qt_inspection_columns.sql
@@ -3,5 +3,5 @@
 -- ============================================
 
 -- Inspection 实体类包含 is_exception 和 exception_reason，但 V4 建表时未创建
-ALTER TABLE qt_inspection ADD COLUMN is_exception INTEGER DEFAULT 0;
+ALTER TABLE qt_inspection ADD COLUMN is_exception BIGINT DEFAULT 0;
 ALTER TABLE qt_inspection ADD COLUMN exception_reason VARCHAR(500);
