+++ dms-app/src/main/resources/db/migration/V7__add_current_scheme_id.sql
@@ -2,4 +2,4 @@
 -- V7: 补全 eq_device 当前运行方案字段
 -- ============================================
 
-ALTER TABLE eq_device ADD COLUMN current_scheme_id INTEGER DEFAULT NULL;
+ALTER TABLE eq_device ADD COLUMN current_scheme_id BIGINT DEFAULT NULL;
